Modernize the PDF code. More CF_OPTIONS/CF_ENUM for XADMaster code. Use the Security framework for decoding PDF AES.

+++ XADMaster/PDF/PDFParser.m
@@ -937,8 +937,8 @@ static BOOL IsDelimiter(uint8_t c);
 		else if([obj isKindOfClass:[NSArray class]])
 		{
 			NSMutableArray *array=obj;
-			int count=[array count];
-			for(int i=0;i<count;i++)
+			NSInteger count=[array count];
+			for(NSInteger i=0;i<count;i++)
 			{
 				id value=array[i];
 				if([value isKindOfClass:[PDFObjectReference class]])
@@ -967,10 +967,10 @@ static BOOL IsDelimiter(uint8_t c);
 		start=[fh readDataOfLength:100];
 	}
 
-	int length=[start length];
+	NSInteger length=[start length];
 	const uint8_t *bytes=[start bytes];
-	int skip=0;
-	for(int i=0;i<length;i++) if(bytes[i]=='\n'||bytes[i]=='\r') skip=i+1;
+	NSInteger skip=0;
+	for(NSInteger i=0;i<length;i++) if(bytes[i]=='\n'||bytes[i]=='\r') skip=i+1;
 	NSString *startstr=[[[NSString alloc] initWithBytes:bytes+skip length:length-skip encoding:NSISOLatin1StringEncoding] autorelease];
 
 	NSData *end=[fh readDataOfLengthAtMost:100];
@@ -1045,7 +1045,7 @@ static BOOL IsDelimiter(uint8_t c);
 	return [other isKindOfClass:[PDFString class]]&&[data isEqual:((PDFString *)other)->data];
 }
 
--(unsigned)hash { return [data hash]; }
+-(NSUInteger)hash { return [data hash]; }
 
 -(id)copyWithZone:(NSZone *)zone
 {
